folder structure and contribution in cpp
