--- Bubble.java
@@ -1,39 +1,0 @@
-import java.util.*;
-public class Bubble
-{
-	public static void main(String args[])
-	{
-		Scanner sc=new Scanner (System.in);
-		System.out.println("Enter the Size of the Array");
-		int n=sc.nextInt();
-		int a[]=new int[n];
-		System.out.println("Enter the Values of the Array");
-		for(int i=0;i<n;i++)
-		{
-			a[i]=sc.nextInt();
-		}
-		System.out.print("Unsorted Array=> ");
-		for(int i=0;i<n;i++)
-		{
-			System.out.print(a[i]+" ");
-		}
-		System.out.println();
-		for(int i=0;i<n-1;i++)
-		{
-			for(int j=0;j<n-1-i;j++)
-			{
-				if(a[j]>a[j+1])
-				{
-					int t=a[j];
-					a[j]=a[j+1];
-					a[j+1]=t;
-				}
-			}
-		}
-		System.out.print("Sorted Array=> ");
-		for(int i=0;i<n;i++)
-		{
-			System.out.print(a[i]+" ");
-		}
-	}
-}
